requirements.txt
requests==2.32.3
pytesseract>=0.3.10
Pillow>=10.0.0
SpeechRecognition>=3.10.0
# pyaudio>=0.2.11  # Install separately: brew install portaudio && pip install pyaudio